Added Early Abandoning + Lower bound DTW
On a 7 million subsequence data top motif takes around 0.18 seconds. But this time grows exponentially as we search for Top K motifs,

The lower bound becomes loose quite fast for top K motifs, also in that case early abandoning slows down the processing, because many distances have to be computed till last. Also this is yet to be tested on billion size dataset.