Removed subsequence spanning across silence region
There were many subsquences which span across silence regions. We want to retain such segments only when the duration of silence is small as some phrases have silence in the middle and pitch extraction is not perfect. Now we have a threshold for that and we remove all subsequences which are more than this threshold. This saves time and space.

Also added start time and end time storage of every motif as because of blacklisting time stamps are not continually stored.