Major changes: Multiple refinement stages
For the experiments we need to compare multiple rank refinment methods. Since these are just rank refinement the top K motifs obtained by the basic DTW version (which is same as used in discovery) essentially is exact for all the rank refinement methods. So since that is the part which takes most of the time, it doesn't make sense to re run experiment changing the rank refinement method alone. Therefore now we can run in one experiment all N number of rank refinement methods (given as input array) and it doesn't take a lot more time than just running one method. So essentially everything (searchFileName etc) becomes a array now and everything is flexible now.