DEBUG: big conceptual problem solved
I was getting repeated motifs from the same file while searching besides all the conditions in the priority queue update to avoid such cases. The problem was because this whole idea of updating priority queue if there is an overlap with a previously stored element works well when we scan the search space only once from time 0 to time end. But because we generated 5 versions of seed motifs for taking distance, we were traversing 5 times the entire search space while using the same priority queue (as the seed motif was the same). This causes many issues. Imagine that we got non overlapping search patterns from the first seed scan(search) in the priority queue and now while going for the second seed version, we updated them and becuase of this two enteries in priority queue are non overlapping. Now when another good pattern comes it finds overlapp with both of them (note that if we scan once such cases do not exist that a new pattern find overlap with 2 enteries because all enteries are non overlapping). Becuase of this we can only replace on entry and thus we have a overlapping entry which could also be exactly the same pattern in 2 different entries.

To solve this I just changed the order of loop. We scan search space incrementally and the seed pattern loop is inside this loop. This saves our ass.