Important change for Carnatic database
Because the annotations are done for each raga, it would be inappropriate to evaluate the result by searching the repetition occurences in other ragas too. Across ragas some patterns might have reps but its virtually impossible to annotate them. Hence a pattern should be searched within the raga. To implement this the patterns are searched only within the files specified in the flist file of the query file.