Major changes: Deleted many files
All the method variants for which I duplicated the code in different files are now embedded in a single file. One can now choose a variant based on the input config file and there is no need to compile diff code for that. Variants which are incorporated in the main code are

1) var 1 and var 2 of the method (diff ways to segment the contours)
2) Different complexity measures
3) Different distance normalizations
4) pa Sa normalization of the melody