Added LSA dimensionality reduction
Wanted to try if LSA type of thing, which is suppose to exploit relationshipt between words (synonyms) etc for document classification helps us. But looks like it does not. This might be because the bottleneck now is the documents which do not have words from the given vocabulary. they have words which are not found in any other document. And so there coouldn't have been any co-occurrence that LSA would have expoited to model or group similar words.