Implemented silence region blacklist differently
Currently removal of segments with big silence regions was done based on time diff between the start and the end of the subseq. But because of compressed flat notes that cant be done (as compressed note also have time elapsed) To handle this subsequences containing silence regions are blacklisted differently.