Debugging/enhancement
I never tried my code with histograms which have sparse notes. As a result of which in some cases there was no valley detected (eg. single note). Since we define minimas as valleys and peak picking depends on minimas this was raising problems for several audio files in Hindustani lessons (Riyaz app)