Enhancement: Snapping of segmentation poit imrpoved
For nyas segmentation we snap the boundaries of the nyas with possible segmentation points. The logic of the snapping point till now was very simple. The starting snaps the the point after that and the ending before that. Sometimes what happens in that we have a possible segmentation point very close to the segmentation point of nyas, but due to temporal constraints it was not picked. Now every segmentation point within a vicinity of 2 samples is picked otherwise temporally as was done earlier.